+++ tests/Console/Application/Theme/DefaultTheme/StateRunningIconTest.php
@@ -7,6 +7,11 @@ namespace Steevanb\ParallelProcess\Tests\Console\Application\Theme\DefaultTheme;
 use PHPUnit\Framework\TestCase;
 use Steevanb\ParallelProcess\Console\Application\Theme\DefaultTheme;
 
+/**
+ * @covers \Steevanb\ParallelProcess\Console\Application\Theme\DefaultTheme::__construct
+ * @covers \Steevanb\ParallelProcess\Console\Application\Theme\DefaultTheme::setStateRunningIcon
+ * @covers \Steevanb\ParallelProcess\Console\Application\Theme\DefaultTheme::getStateRunningIcon
+ */
 final class StateRunningIconTest extends TestCase
 {
     public function testDefaultValue(): void