+++ tests/Process/Process/CanceledOutputVerbosityTest.php
@@ -7,6 +7,11 @@ namespace Steevanb\ParallelProcess\Tests\Process\Process;
 use PHPUnit\Framework\TestCase;
 use Steevanb\ParallelProcess\Tests\CreateLsProcessTrait;
 
+/**
+ * @covers \Steevanb\ParallelProcess\Process\Process::__construct
+ * @covers \Steevanb\ParallelProcess\Process\Process::setCanceledOutputVerbosity
+ * @covers \Steevanb\ParallelProcess\Process\Process::getCanceledOutputVerbosity
+ */
 final class CanceledOutputVerbosityTest extends TestCase
 {
     use CreateLsProcessTrait;