+++ tests/Process/Process/NameTest.php
@@ -7,6 +7,11 @@ namespace Steevanb\ParallelProcess\Tests\Process\Process;
 use PHPUnit\Framework\TestCase;
 use Steevanb\ParallelProcess\Tests\CreateLsProcessTrait;
 
+/**
+ * @covers \Steevanb\ParallelProcess\Process\Process::__construct
+ * @covers \Steevanb\ParallelProcess\Process\Process::setName
+ * @covers \Steevanb\ParallelProcess\Process\Process::getName
+ */
 final class NameTest extends TestCase
 {
     use CreateLsProcessTrait;