+++ changelog.md
@@ -1,5 +1,7 @@
 ### master
 
+### [0.3.0](../../compare/0.2.1...0.3.0) - 2022-01-10
+
 - Create GitHub Actions to build and push Docker image `steevanb/php-parallel-processes:ci`
 - Create GitHub Actions to build and push Docker image `steevanb/php-parallel-processes:parallel-processes-x.y.z` and automatically call it on new release
 - Create GitHub Actions to build and push Docker image `steevanb/php-parallel-processes:release`